Requirements, readme, and data loader
Just to start things off, a requirements.txt has been added as well as
some general information on how to setup the virtualenv (pip) in the
readme. In addition, a simple data loading script has been added; the
script probably won't be used as it is, but it will serve as a nice
reminder on the format that the scripts should take from now on.